Hide errors for removing files that don't yet exist but should
During a fresh installation of the dotfiles the setup script will
attempt to remove files that are to be symlinked that might not exist
yet. This makes `rm` throw a fit and display error messages on screen
which is meh.

This commit sends those potential error messages to /dev/null.