Add Vagrantfile to dotfiles
I find this particular Vagrantfile very useful and needed to save it
somewhere. I use it mainly when testing out PE with various setups but
it could be used and adapted for many other setups.

I'm going to start putting more of my scripts and useful files that do
not need to be on the path in here for syncing between my computers and
sharing with others. I'll have another structure hopefully soon on how
to better organize these files like symlinking the files directory to
~/projects/misc_scripts or something of the like. And potentially making
it so that it doesn't block the update script as there will probably be
a lot of WIP scripts in this directory.