Disable nixd in neovim
This dependency has a CVE associated with it and is no longer
installable without an override. We are disabling this for now until the
package can be updated in Nix.

This also has fallen out of the package cache which requires nixd to be
compiled from scratch, a costly event.