feat: Add terraform_env and update set_infra
This commit allows set_infra to be overridden with either the first
argument as a tailscale IP. Passing an IP into this function will
utilize consul running on that node to get the rest of the
infrastructure IP addresses for the environment.

This commit also adds in my `terraform_env` helper which loads
environment variables from the infrastructure kv store within Vault in
order for terraform to have all of the environment variables needed to
run properly against my environment.