feat: Add more resilient set_infra
This commit enables a more resilient set_infra by automatically
detecting if consul can be reached on the current network. If it cannot
tailscale will be queried for an active consul node in which case the
rest of the infrastructure will be automatically discovered. A tailscale
ip or hostname can be passed into set_infra to force discovery on that
particular node.